Add files via upload
"This is a question I received from Codecademy to practice my command line skills. The task involves finding all files with the .txt extension in continent directories, checking if they are empty, and listing the names of the empty ones. The result is then saved in a file named empty_files.txt located in the todo directory."